# fern/openapi/openapi.yml
...
openapi: 3.1.0
info:
  title: Navigator AI
  description: Live interactive demo agent for any web product.
  version: 0.1.0
paths:
  /v1/products:
    post:
      summary: Register Product
      description: Register a product. The API key in the response is shown exactly once.
      operationId: register_product_v1_products_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProduct'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisteredProduct'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/products/me:
    get:
      summary: Whoami
      operationId: whoami_v1_products_me_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/products/site-graph:
    put:
      summary: Upload Site Graph
      description: 'Validate and store a new site graph revision, and make it active.


        Validation runs before anything is written, so a rejected upload cannot break

        a live demo. The error text is identical to what the file loader produces --

        there is one validator in the system, deliberately.'
      operationId: upload_site_graph_v1_products_site_graph_put
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteGraphUpload'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteGraphRevision'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Get Site Graph
      operationId: get_site_graph_v1_products_site_graph_get
      parameters:
      - name: revision
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Revision
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteGraphRevision'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/products/site-graph/revisions:
    get:
      summary: List Revisions
      operationId: list_revisions_v1_products_site_graph_revisions_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SiteGraphRevision'
                title: Response List Revisions V1 Products Site Graph Revisions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/products/site-graph/activate:
    post:
      summary: Activate Revision
      description: Roll back to an earlier revision.
      operationId: activate_revision_v1_products_site_graph_activate_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_activate_revision_v1_products_site_graph_activate_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/products/flows:
    get:
      summary: List Flows
      description: Which flows this product's active site graph offers, per page.
      operationId: list_flows_v1_products_flows_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
                title: Response List Flows V1 Products Flows Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/demos:
    post:
      summary: Start Demo
      description: Start a demo. Returns immediately; poll GET /v1/demos/{id} for progress.
      operationId: start_demo_v1_demos_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewDemo'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: List Demos
      operationId: list_demos_v1_demos_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DemoView'
                title: Response List Demos V1 Demos Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/demos/start:
    post:
      summary: Start Live Demo
      description: 'Create a meeting for *this* session and put Navigator in it, now.


        The meeting is instant, not scheduled: it exists and is joinable the moment

        this returns, which is the only useful semantics for a "Show Demo" button.

        The link is minted per call, so two prospects can be demoed at once and no

        human has to set NAVIGATOR_MEETING_URL first. The response carries the join

        URL, which is what the button redirects to.


        Ordering is deliberate: the graph is loaded and the flow validated *before*

        any meeting is created, so a bad request never leaves an orphaned meeting

        behind.'
      operationId: start_live_demo_v1_demos_start_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartLiveDemo'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveDemoView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/demos/{demo_id}:
    get:
      summary: Get Demo
      operationId: get_demo_v1_demos__demo_id__get
      parameters:
      - name: demo_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Demo Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/demos/{demo_id}/end:
    post:
      summary: End Demo
      operationId: end_demo_v1_demos__demo_id__end_post
      parameters:
      - name: demo_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Demo Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/demos/{demo_id}/actions:
    get:
      summary: Demo Actions
      description: 'The ActionLog for one demo: every call, its expectation, and what happened.


        This is the audit trail that makes the product sellable -- a customer can see

        exactly what the agent did in front of their prospect.'
      operationId: demo_actions_v1_demos__demo_id__actions_get
      parameters:
      - name: demo_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Demo Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionLogEntry'
                title: Response Demo Actions V1 Demos  Demo Id  Actions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/products/failures:
    get:
      summary: Product Failures
      description: Recent failures across all demos of this product -- which flows are rotting.
      operationId: product_failures_v1_products_failures_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionLogEntry'
                title: Response Product Failures V1 Products Failures Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/products/corrections/pending:
    get:
      summary: Pending Corrections
      description: 'Reflection output awaiting human approval.


        Never auto-promoted: an agent that can silently rewrite its own rules is

        not debuggable.'
      operationId: pending_corrections_v1_products_corrections_pending_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Pending Corrections V1 Products Corrections Pending Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/products/corrections/{correction_id}/approve:
    post:
      summary: Approve Correction
      description: "Human approves a pending rule \u2192 live Chroma corrections collection."
      operationId: approve_correction_v1_products_corrections__correction_id__approve_post
      parameters:
      - name: correction_id
        in: path
        required: true
        schema:
          type: string
          title: Correction Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/ApproveCorrectionBody'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Approve Correction V1 Products Corrections  Correction Id  Approve Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/products/corrections/{correction_id}/reject:
    post:
      summary: Reject Correction
      operationId: reject_correction_v1_products_corrections__correction_id__reject_post
      parameters:
      - name: correction_id
        in: path
        required: true
        schema:
          type: string
          title: Correction Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reject Correction V1 Products Corrections  Correction Id  Reject Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NewProduct:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        product_id:
          anyOf:
          - type: string
            pattern: ^[a-z0-9][a-z0-9_-]{1,63}$
          - type: 'null'
          title: Product Id
      type: object
      required:
      - name
      title: NewProduct
...
    Product:
      properties:
        product_id:
          type: string
          title: Product Id
        name:
          type: string
          title: Name
        created_at:
          type: string
          format: date-time
          title: Created At
        active_revision:
          anyOf:
          - type: integer
          - type: 'null'
          title: Active Revision
      type: object
      required:
      - product_id
      - name
      - created_at
      title: Product
    RegisteredProduct:
      properties:
        product:
          $ref: '#/components/schemas/Product'
        api_key:
          type: string
          title: Api Key
      type: object
      required:
      - product
      - api_key
      title: RegisteredProduct
      description: Registration response. The API key is shown exactly once.
    SiteGraphRevision:
      properties:
        product_id:
          type: string
          title: Product Id
        revision:
          type: integer
          title: Revision
        source:
          type: string
          enum:
          - yaml
          - recorded
          - sdk
          title: Source
        yaml:
          type: string
          title: Yaml
        created_at:
          type: string
          format: date-time
          title: Created At
        site:
          type: string
          title: Site
        graph_version:
          type: integer
          title: Graph Version
      type: object
      required:
      - product_id
      - revision
      - source
      - yaml
      - created_at
      - site
      - graph_version
      title: SiteGraphRevision
    SiteGraphUpload:
      properties:
        yaml:
          type: string
          minLength: 1
          title: Yaml
        source:
          type: string
          enum:
          - yaml
          - recorded
          - sdk
          title: Source
          default: yaml
      type: object
      required:
      - yaml
      title: SiteGraphUpload
...